refactor(ui-voip): extract PeerCardsView from MediaCallRoomSection
