+++ .changeset/extract-peer-cards-view.md
@@ -1,0 +1,5 @@
+---
+"@rocket.chat/ui-voip": patch
+---
+
+Extract the media call peer cards rendering from `MediaCallRoomSection` into a dedicated `PeerCardsView` component. Pure refactor, no behavior change.
