+++ packages/ui-voip/src/views/MediaCallRoomSection/MediaCallRoomSection.tsx
@@ -1,16 +1,13 @@
-import { Box, Button, ButtonGroup } from '@rocket.chat/fuselage';
-import { memo, useState } from 'react';
+import { Box, ButtonGroup } from '@rocket.chat/fuselage';
+import { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
+import PeerCardsView from './PeerCardsView';
 import {
 	ToggleButton,
 	Timer,
 	DevicePicker,
 	ActionButton,
-	CardListContainer,
-	CardListSection,
-	PeerCard,
-	StreamCard,
 	useShouldWrapCards,
 	CARD_LIST_SECTION_MAX_HEIGHT,
 	ActionStrip,
@@ -19,7 +16,6 @@ import {
 import { useMediaCallInstance } from '../../context/MediaCallInstanceContext';
 import { useMediaCallView } from '../../context/MediaCallViewContext';
 import useRegisterView from '../../context/useRegisterView';
-import { usePlayMediaStream } from '../../providers/usePlayMediaStream';
 
 type MediaCallRoomSectionProps = {
 	showChat: boolean;
@@ -48,7 +44,6 @@ const getSplitStyles = (showChat?: boolean) => {
 const MediaCallRoomSection = ({ showChat, onToggleChat, user, containerHeight }: MediaCallRoomSectionProps) => {
 	const { t } = useTranslation();
 
-	const [focusedCard, setFocusedCard] = useState<'remote' | 'local' | null>('remote');
 	const {
 		sessionState,
 		onMute,
@@ -58,73 +53,24 @@ const MediaCallRoomSection = ({ showChat, onToggleChat, user, containerHeight }:
 		onToggleScreenSharing,
 		onOpenPopout,
 		onClosePopout,
-		streams: { remoteScreen, localScreen },
+		streams: { localScreen },
 	} = useMediaCallView();
 	const { currentViews } = useMediaCallInstance();
 
 	const isPopout = currentViews.includes('popout');
 
-	const { muted, held, remoteMuted, remoteHeld, peerInfo, connectionState, startedAt } = sessionState;
+	const { muted, held, peerInfo, connectionState, startedAt } = sessionState;
 
 	const shouldWrapCards = useShouldWrapCards(showChat, containerHeight);
 
 	const connecting = connectionState === 'CONNECTING';
 	const reconnecting = connectionState === 'RECONNECTING';
 
-	const [remoteStreamRefCallback] = usePlayMediaStream(remoteScreen?.stream ?? null);
-	const [localStreamRefCallback] = usePlayMediaStream(localScreen?.stream ?? null);
-
 	useRegisterView('room');
-
-	const onClickFocusRemoteCard = () => {
-		setFocusedCard((prev) => (prev === 'remote' ? null : 'remote'));
-	};
-
-	const onClickFocusLocalCard = () => {
-		setFocusedCard((prev) => (prev === 'local' ? null : 'local'));
-	};
 
 	if (!peerInfo || 'number' in peerInfo) {
 		return null;
 	}
-
-	const remoteStreamCard = remoteScreen?.active ? (
-		<StreamCard onClickFocusStream={onClickFocusRemoteCard} focused={focusedCard === 'remote'}>
-			<video
-				preload='metadata'
-				style={{ objectFit: 'contain', height: '100%', width: '100%' }}
-				ref={remoteStreamRefCallback}
-				autoPlay={true}
-				muted={true}
-				playsInline={true}
-			>
-				<track kind='captions' />
-			</video>
-		</StreamCard>
-	) : null;
-
-	const localStreamCard = localScreen?.active ? (
-		<StreamCard
-			own
-			onClickFocusStream={onClickFocusLocalCard}
-			onClickStopSharing={onToggleScreenSharing}
-			focused={focusedCard === 'local'}
-			showStopSharingOnHover
-		>
-			<video
-				preload='metadata'
-				style={{ objectFit: 'contain', height: '100%', width: '100%' }}
-				ref={localStreamRefCallback}
-				autoPlay={true}
-				playsInline={true}
-				muted={true}
-			>
-				<track kind='captions' />
-			</video>
-		</StreamCard>
-	) : null;
-
-	const focusedCardElement = focusedCard === 'remote' ? remoteStreamCard : localStreamCard;
 
 	return (
 		<Box
@@ -138,26 +84,7 @@ const MediaCallRoomSection = ({ showChat, onToggleChat, user, containerHeight }:
 			aria-label={t('Voice_call')}
 			{...getSplitStyles(showChat)}
 		>
-			<CardListSection>
-				{isPopout && (
-					<Box mb={20} p={24} w='full' display='flex' flexDirection='column' justifyContent='space-between' alignItems='center'>
-						<Box is='h1' color='font-default' mbe={40}>
-							{t('Call_open_separate_window')}
-						</Box>
-						<Button onClick={onClosePopout} icon='arrow-from-cross-box' large>
-							{t('Show_call_here')}
-						</Button>
-					</Box>
-				)}
-				{!isPopout && (
-					<CardListContainer focusedCard={focusedCard ? focusedCardElement : undefined} shouldWrapCards={shouldWrapCards}>
-						<PeerCard displayName={user.displayName} avatarUrl={user.avatarUrl} muted={muted} held={held} />
-						<PeerCard displayName={peerInfo.displayName} avatarUrl={peerInfo.avatarUrl} muted={remoteMuted} held={remoteHeld} />
-						{focusedCard !== 'remote' && remoteStreamCard}
-						{focusedCard !== 'local' && localStreamCard}
-					</CardListContainer>
-				)}
-			</CardListSection>
+			<PeerCardsView user={user} shouldWrapCards={shouldWrapCards} />
 			<ActionStrip
 				leftSlot={
 					<Box color='default' alignContent='center' pis={16}>
